pyproject.toml
[tool.poetry]
name = "banappeals"
version = "0.1.0"
description = ""
authors = ["Snaacky <snaacky@pm.me>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-Discord = "^0.1.69"
dataset = "^1.5.2"
gunicorn = "^20.1.0"
python-geoip-python3 = "^1.3"
python-geoip-geolite2 = "^2015.0303"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
